+++ CHANGELOG.md
@@ -5,7 +5,8 @@ This file is a history of the changes made to idearium-lib.
 ## Unreleased
 
 - Added `Logger` instance to the common folder.
-- Added new `util` library.
+- Added new `util` to the library.
+- Added new `mongodb` to the common folder.
 
 ## 1.0.0-alpha.11