+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This file is a history of the changes made to idearium-lib.
 
-## Unreleased
+## 1.0.0-alpha.12
 
 - Added `Logger` instance to the common folder.
 - Added new `util` to the library.